requirements.txt
flask
python-dotenv
python-slugify